BUG/MAJOR: JavaScript errors due to localStorage changes
History and bookmark queries are stored locally using the localStorage
browser API. The recent global code REORG has changed the format of
the data stored, a property was renamed from query to queryInfo.

That change leads to javascript errors and various parts of the interface
do not work, for instance menu buttons.

Since this breaks compatibility, this patch gets back to the original
property name. It should not be changed until a proper cleanup or rewrite
of the localStorage related code is done.