BUG/MEDIUM: autotable: Fix app error when no items
The App throws a bunch of errors if the array of tableItems returned by the
API is null and can't be iterated. Also made fetchTableItems async to allow
the app to wait for the data from the API before handling the response, thus
removing additional potential errors.

EDIT JTH:
fetchTableItems() to review in depth:
- array_data props, unArray(path) !== '' condition
- tableData === [] is most probably wrong
- async / await to clarify since the response is handled in then()
- test all on error cases, request error, empty data, broken array_path, etc.