MAJOR: Rework the whole navigation and data persistence across the app
This change completely reworks the way the data is shared across the whole
app, between components, routes, URLs, and the persistence of this data.
Most of the dynamic data of the app was moved from the Vuex Store to the Route
query params, to be able keep data persistent on URL sharing, bookmarking,
opening on different tabs, devices etc. It is then closely watched by VueJS
in case of changes and Views are updated accordingly. This happened to fixed
a number of bugs regarding the search history and bookmarking, as well as the
Timeline always saving the state of its selected database now.