MINOR: host: Replace axios.all() and axios.spread() with ES6 equivalent
Methods axios.all() and axios.spread() are deprecated and can be replaced
by standard ES6 code.

Note, here the catch() block will never be executed as expected because
the Api.get() axios wrapper is used and its error handling code is wrong.
Its purpose was to catch Api errors in order to display messages from a
central point, but since the error is never rethrown, the promise always
resolves. This will be addressed in another commit.