BUG/MAJOR: api: Rework error handling, deprecate get()
get() provides error handling code to render axios error messages
automatically in a generic way, from a central point. This is a
good idea. However, since the error is never re-thrown, the promise
always resolves, even if the request failed, in which case undefined
is returned as the resolved (success) value -- undefined because the
catch() in get() returns the result of handleError(), ie. undefined
since there is no return statement in that function.

It makes the then() / catch() method useless because 1) the catch is
never called on an axios error, 2) every result should be checked for
undefined in the then() callback in order to handle the error case,
which is the whole point of catch(). The same reasoning applies to the
try { await ... } catch () { ... } method.

Since the whole purpose of catch'ing in the get() wrapper was to
notify errors from a central point, let's use instead Axios builtin
interceptors.

The patch introduces a new helper exposed by the api plugin to make
requests: axiosData(). Eventually we expect all code in the project
to migrate to this new function, which is why get() got the
@deprecated flag. Once code is migrated, get() will be removed, along
with its dependency function handleError().

axiosData() is not defined async. There is no point in await'ing the
promise returned by axios like it's done in get(). It is up to the
caller to await for a result, or to use the callback-based then()
chaining method.

axiosData(), is mostly equivalent to get() in that it resolves with
Axios response.data and emits an error message on failure. The main
difference is that it still rejects on error, that way the then() /
catch() syntax still have sense. Note, the automatic error notification
can be disabled, allowing the caller to handle the whole error if he
wants.

axiosData('http://...')
    .then((mydata) => { /* play with returned data */ })
    .catch(() => { /* already notified, anything else? */ });

let mydata;
try { mydata = await axiosData('http://...'); /* play... */ }
catch(e) { /* already notified, anything else? */ }

Another point I found annoying in get(), yet not important in this
project so far anyway, it always returns the axios response.data.
This does not allow the caller to check nor use any other information
available in the axios response object: HTTP status code, response
headers, etc. Since this kind of information is not needed here so
far, we kept the original behavior in the axiosData() replacement.
The "data" in "axiosData" is there on purpose to make it clear it
resolves with Axios response.data.