BUG/MINOR: init: Rework main init so we can notify early errors
The patch reworks the initialization of the application in order
to make it possible to notify the user of early errors. Such errors
are related to axios requests that take place before Vue gets
instanciated. These errors now get buffered, then rendered using
the new Message component once the Vue app is mounted.