MEDIUM: main: Get menu definition out of the main view component
The main app menu is now generated automatically from the routes
definition. Let's update ViewMain to get that part out of if.

We want to centralize all the configuration related to the routes
at the same place. The same approach will be taken in a next commit
to enable or disable the timeline depending on the route / view.