MAJOR: main: Rework recent queries and related local storage
The code related to the local storage of the recent queries (search
history or bookmark) was difficult to maintain. It would likely
cause exceptions if the format of the data stored in the browser was
to change.

The interface with the localStorage object is now out of the main
view component, it was moved to the store plugin. A format version
number is now attached to the localStorage data. It is expected to
match the one defined in LOCAL_STORAGE_VERSION. This allows for
automatic migration if the format of the data were to be changed: we
would then increment LOCAL_STORAGE_VERSION and implement the migration
code in migrateLocalStorage().

This rework introduces two new function properties on the routes meta
objects: buildHistory() and buildBookmark(). Those are called in
order to produce history and bookmark records. Both return an object
with a "label" property to display in the interface, and an "entry"
property, which is a partial route to replay the query.