MINOR: vlan: Keep small z-index values for the fixed headers internals
It is better to keep small z-index values within internals of a specific
components, so they don't go over other more general parts related to
the layout of the interface. The side panel (navigation drawer) has a
default z-index value of 6, so staying below that value here will avoid
to increase it later.