MINOR: vlan: Add right border on the first column header
When scrolling horizontally on the vlan matrix table, the first column
containing the name of the switches remains fixed, it is a header. There
is now a right border when scrolling on that column. As a consequence,
the left border of the second column has been removed in order to prevent
having tow borders when horizontal scrolling is not in effect.