MINOR: timeline: Adjust x-axis and tooltip time formats
On the x-axis, hours were displayed on a 12-hours time format without
any a.m. or p.m. suffix. This was confusing so let's use a 24-hours
format instead.

The commit also changes the date format displayed in the tooltip when
hovering a data point. The new format includes a timezone offset so
we do not wonder anymore if the date is in UTC or in localtime.