MAJOR: timeline: Review the whole logic to make it a generic component
The component is now more generic and can be used as a standalone
timetick selector. This a major rewrite, removing lots of unnecessary
complexity due to dependencies there were on the store (available
databases) and the route query parameters (selected database).

The ticks (available databases) are now passed with a :databases
property. Similarly, the value to emphasize (selected database) is
now passed with a !value property. The selection of point or tick
(new database) now triggers a @click event carrying the selected value.
It is up to the parent component to handle the @click event and to
update the TimeLine :value property in order to make the change
visible.