MINOR: timeline: Display a message when the is no database available
When mounting the TimeLine component, a <div> get appended to the main
element. It's purpose is to to display a "This is no database available!"
message. Positioning is done with CSS class "timeline-no-data" and
vuetify CSS class "d-none" is used to toggle display of the <div>
depending on the data.