BUG/MINOR: Entity selector null value change after clearing text input
The autocomplete components emit a @change event with a null value
when the text input is cleared. This null value was not ignored,
leading to a redirection on the initial entity-picker page.

Once this was addressed, we hit another annoying behavior. Normally,
after loosing focus on the autocomplete text input without selecting a
new value, the current value gets restored in the text input. This is
not happening if the text input is cleared.

- Proper behavior. Focus on the text input. Delete a few characters,
  not all of them. Then, loose focus from the text input. The value of
  the text input get restored to the current value of the selector.

- Annoying behavior. Same as above, but clear the text input completely.
  The current value of the selector value does NOT gets restored in the
  text input.

To fix it, a @blur event handler was added. We just force a refresh of
the entity selector by updating its :key.

The v-model of the entity selector was removed in this patch. Its use
did not give any advantage here because changing the selected value is
already processed in a @change handler. The value to preset in the list
is now given with the :value property.