MAJOR: switch: Review the whole logic handling state change
The main focus is "do NOT use async/await here". It's called from
an immediate watch that reacts on query parameters, I just perfer
to avoid that in this context of a Vue app!

Big changes made previously to handle context changes allow to
simplify corner cases. On entity change, we redirect to the main
view and skip if any of the required parameters are empty. I like
it that the module itself decides of the action to do on entity
change, this is module specific.

Note we also skip if values in apiStateParams do not change.
Removing it would be totally fine, we may do it in future. It would
allow to force refresh the data from the interface, eg. (re)click
on the selected timeline point.