BUG/MINOR: vlan: Delay initial table height update to next tick
The issue appeared after first batch of major reworks that were made
following v1.0.2. On full page reload, the first time the table height
gets computed, it is too tall by 48px.

It was tested that delaying the initial table hieght computation to
nextTick fixes the issue. This is what is done here.

If it turns out the issue comes back in future, it was also tested
that triggering a (re)computation in the updated() hook fixes the
issue as well. So for reference it would be as follows:

updated() {
    this.setTableHeight();
}