FEATURE: app: Emit an event for shortcut ctrl-shift-f
Shortcut Ctrl+Shift+F will be used to trigger focus on the search menu.
The reasoning to put the window keydown handler in App is that the
application notifies potential listeners, and those implement an action
depending on the target (SearchMenu, VTextField, VAutoComplete, etc)
and the view (eg. entity selector, main app wrap).