MEDIUM: api: Let init() throw so we get a chance to notify the user
Let init() throw on axios error or if the response has unexpected data.
The error is ment to be catched on the caller side, giving it a chance
to notify the user once Vue is mounted.

This commit is tagged as MEDIUM because it leaves a promise without
rejection handling when the api plugin gets initialized, in main.js.
This will be addressed in a separate commit.