MAJOR: main: Review the whole logic leading to app context change
Context change is related to route query parameters for the
current entity and database.

In the context of a Vue application, avoid using async/await,
especially with immediate watches on $route.query and when there
are redirections in play updating the query parameters.

Keep in mind that immediate watches are run at created() time
of the component, not at mounted time. In case of async handlers,
you may await in the function, but Vue - the caller - does not
await the promise returned implicitly by an async function. Unless
you just know what you do, this could lead to strange behaviors.

The principle we stick with here is: user interaction updates the
query parameters, on which we react to update the context and data
(entity, database).

This patch is the base for more global rework in children views.
In next commits, we will focus on these to remove async from
immediate watches on route query parameters.