FEATURE: router: Generate app menu from routes definition
We make use of the meta object available in routes definition
to attach information used to generate the main app menu.

The result is accessible via a getMenu() method attached to
Vue router, eg. this.$router.getMenu() in components.

This will be used by ViewMain in a next commit so we can get
the menu definition out of it.