FEATURE: search: New SearchMenu component for global search accross the app
This component will be used in a next commit to replace the search
input in the top panel of the application. The goal is to introduce
global search accross multiple modules with an approach similar to
the search menu on GitHub.

The configuration of the component is based on property :entries,
an array providing a list of possible search targets (partial routes)
with display options (icon, label, flag to indicate a possible empty
search).

The :value property is used to pass the current search text to
preset in the search input.

After user interaction, when a new search is to be triggered, the
component emits a @submit event carrying the necessary data for the
listener to process the search. It is up to the parent component to
handle the @submit event and to update the SearchMenu :value property
in order to make the change visible in the search input.