FEATURE: router: Generate search menu from routes definition
We make use of the meta object available in routes definition
to generate definition for possible targets of the future global
search menu.

The result is accessible via a getSearchMenu() method attached
to Vue router, eg. this.$router.getSearchMenu() in components.

This will be used by ViewMain in a next commit to replace the
current search input by the new SearchMenu component.