MAJOR: fdb: Review the whole logic handling state change
Big changes made previously to handle context changes allow to
simplify the code in apiStateParams watch. During the transition
to a new entity, the database value first gets removed in query
parameters. This is done from ViewMain. We only need to skip if any
of the required parameters are empty.

Note we also skip if values in apiStateParams do not change.
Removing it would be totally fine, we may do it in future. It would
allow to force refresh the data from the interface, eg. (re)submit
the search, (re)click on the selected timeline point.