MINOR: inventory: Prepare for multiple icons in flags column
We want to be able to render multiple flag icons for a device. This
patch prepares for it. Icons can now be accumulated in an array, and
join of them is returned.

This commit includes some coding style cleanup on the lines affected
by the patch: avoid too long lines and inline "if".