MINOR: search: Loose input focus on v-text-field blur
Previously we were loosing focus on submit. That was not the most
convenient when you have to refine your search, you need to focus
on the search input to type again (eg: mouse click, keyboard
shortcut).

Originally we wanted to keep focus on submit but annoying behavior
brought us to think it again: focus get (re)triggered (ie. menu opens)
on successive alt-tab, when back to the browser window.

We found another way to fix the alt-tab issue, just force to loose
focus on the search input when the v-text-field emits a blur event.
Let's try it that way.