FEATURE: backup: ttyd integration proof of concept
This is a proof of concept to add a new flag link in the inventory
table. When present the link triggers a ttyd popup that connects and
opens a terminal on the corresponding device.

Due to how the autotable works, we had difficulties to do something
very simple: add a click handler to trigger the popup window. It
highlights that we need to find a proper Vue way to customize autotable
cells content, in that case probably with a slot instead of HTML
generated from a column specific format() function.

This patch is a first shot, it is functional. However the declaration
of the click handler, and the way we pass a parameter to it, are hugly!
This will be addressed in further commits.