BUG/MINOR: timeline: Fix point color when the first database is selected
Pierre Delclaux reported a case where the selected database does not
get highlighted on the timeline.

It turns out this happens when the first database of the timeline is
selected, ie. index 0 in the array. It's a classic stupid mistake, a
comparison was expecting a truthy value, as in "is it defined?", which
obviously does not pass with a zero-value number.