MAJOR: vlan: Alternate view with vlan names on the table header
The vlan ids row was previously defined in the table body. It was moved
to the table header, grouped with the vlan labels on top, and the whole
table header was made sticky.

This patch was originally implemented by Eddy Gangloff, then reworked
to be applied on the current master.