MAJOR: autotable: Review copyable to make it a column option
The copyable column option controls whether or not a clipboard copy
icon get displayed on the right edge of cells when the maouse is over
a non-empty cell.

The result of the rendering (has text content or not) cannot be
known from the template (format text or html, or vue slot), so the
feature is implemented with mouse enter/leave event handlers. We were
a bit reluctant to registered lots of event handlers. We did some
testing and it turns out the performance it is the same as before.