BUG/MEDIUM: inventory: Fix syntax error during build due to style comments
We hit this error when building:
Cannot read properties of undefined (reading 'spaces')

Looks like it affects only ViewInventory. It can be reproduced with
vue-cli, when running vue-cli-service serve (not with vue-cli-service
build) or with vite build.

The issue occurs in a <style> block, when using a block-comment /* ... */
nested under a ::v-deep { ... }. The following solution seems to fix the
issue; this patch using the first:
- changing ::v-deep to *::v-deep
- changing ::v-deep to .<class>::v-deep or #<id>::v-deep and adding
  either .<class> or #<id> to <AutoTable> in the template
- changing block-comments /* ... */ to line comments // ...