MINOR: network: Avoid regexp named capture groups for compatibility
On Firefox 76, the web interface does not load because an exception
is emited due to regexp named capture groups not supported by the
browser (SyntaxError: invalid regexp group).

It will be supported on Firefox 78. Although those are quite old
Firefox releases, since it is not too much hassle, let's patch to
remove the use of named groups.